requirements.txt
opencv-python==4.10.0.84
numpy==1.26.4
scipy==1.13.1
Pillow==11.1.0
matplotlib==3.10.0
scikit-image==0.25.0
mediapipe==0.10.20